+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Added
     - ``get_last_material_index()`` helper (#1554)
     - ``get_latest_file_path()`` helper (#1554)
     - "Not found" element for iRODS modal filter (#1562)
+    - Existing iRODS file check in material name editing (#1494)
 
 Changed
 -------